pyproject.toml
[project]
name = "generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.6"
dependencies = [
    "accelerate==1.13.0",
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==7.0.5",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.5",
    "click==8.3.1",
    "colorama==0.4.6",
    "diffusers==0.37.0",
    "exceptiongroup==1.3.1",
    "filelock==3.25.2",
    "fsspec==2026.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "h11==0.16.0",
    "hf-xet==1.4.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.7.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "jinja2==3.1.6",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "narwhals==2.18.0",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "packaging==26.0",
    "pandas==2.3.3",
    "pillow==12.1.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "pyarrow==23.0.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2026.2.28",
    "requests==2.32.5",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "safetensors==0.7.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.3",
    "streamlit==1.55.0",
    "sympy==1.13.1",
    "tenacity==9.1.4",
    "tokenizers==0.22.2",
    "toml==0.10.2",
    "torch==2.6.0+cu124",
    "torchaudio==2.6.0+cu124",
    "torchvision==0.21.0+cu124",
    "tornado==6.5.5",
    "tqdm==4.67.3",
    "transformers==5.3.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "watchdog==6.0.0",
    "zipp==3.23.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"